+++ Framework/DeltaDNA.h
@@ -38,5 +38,7 @@ FOUNDATION_EXPORT const unsigned char DeltaDNAVersionString[];
 #import <DeltaDNA/DDNAEngagement.h>
 #import <DeltaDNA/DDNAEngageFactory.h>
 #import <DeltaDNA/DDNAImageMessage.h>
+#import <DeltaDNA/DDNAEventAction.h>
+#import <DeltaDNA/DDNAEventActionHandler.h>
 
 #endif /* _DELTADNA_ */